% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_tools.R
\name{rmd2gm}
\alias{rmd2gm}
\title{Converts a music data.frame into a gm music object}
\usage{
rmd2gm(rmd, key = "Am", clef = "G", meter = c(4, 4), unit = 1/8)
}
\arguments{
\item{rmd}{\code{data.frame} input music table in standard \code{rmusic} format}

\item{key}{\code{character} the signature key}
}
\value{
gm object containing the music in gm format
}
\description{
\code{gm} is a R package for working with music.
This function converts a standard rmusic table into a gm object from which
you can generate a note sheet or export to midi format
}
