% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabr_tools.R
\name{rmd2tabr}
\alias{rmd2tabr}
\title{Convert a rmusic data.frame into a tabr music object:}
\usage{
rmd2tabr(rmd, unit = 1/8, ...)
}
\arguments{
\item{rmd}{\code{data.frame} input music table in standard \code{rmusic} format}

\item{key}{\code{character} the signature key}
}
\value{
gm object containing the music in tabr format
}
\description{
\code{tabr} is a R package for working with music.
This function converts a standard rmusic table into a tabr object from which
you can generate a note sheet or export to midi format
}
