% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snote_tools.R
\name{smeasure}
\alias{smeasure}
\title{Extracts a measure number from \code{smelody}}
\usage{
smeasure(smelody, mn)
}
\arguments{
\item{mn}{\code{numeric/integer}: measure number}
}
\description{
Returns single-measure part of a \code{smelody} string associated with the given
measure number from a multi-measure \code{smelody} string
returns a \code{smelody} associated with the given measure number.
}
\examples{
s_melody = '2c3CEEDDCCDCa4bc5bDDCCbbCbg6af7EGGGGEEFEC8DEDb9CgED10DCbCfgf11FAAGGFFGFD12Ef'
s_measure(s_melody, 5)
}
